+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Added optional `$field` parameter to `FlexMediaTrait::deleteMediaFile()` method
     * Added `has_env()` twig function to test if environmental variable is defined
 2. [](#improved)
+    * **BC BREAK** `Medium` no longer extends `Data`
     * By default, add media to only pages which have been initialized in pages loop
 3. [](#bugfix)
     * Fixed locking and reading files using `Framework\File` classes